requirements.txt
# Core
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
bcrypt==4.1.3
python-dotenv==1.0.1
gunicorn==22.0.0

# Integrations
requests==2.32.3
feedparser==6.0.11
boto3==1.34.0
Pillow==10.3.0
bleach==6.1.0
itsdangerous==2.2.0
markdown==3.6

# Excel reading (for seed)
openpyxl==3.1.2
